fully completed the basis of the controller and business layer tests. some tests still need certain debugging.

+++ src/main/java/com/rooshdashboard/rooshdashboard/business/IParkingGarage/DeleteParkingGarageUseCase.java
@@ -3,5 +3,5 @@ package com.rooshdashboard.rooshdashboard.business.IParkingGarage;
 import com.rooshdashboard.rooshdashboard.domain.ParkingGaragee.DeleteParkingGarageResponse;
 
 public interface DeleteParkingGarageUseCase {
-    DeleteParkingGarageResponse deleteParkingGarage(int id);
+    DeleteParkingGarageResponse deleteParkingGarage(Long id);
 }
